Create RubyInstalledAndAvailable goal
